+++ src/classes/leap-year.ts
@@ -1,6 +1,7 @@
 import {LeapYearRules} from "../constants";
 import ConfigurationItemBase from "./configuration-item-base";
 import {LeapYearConfig, LeapYearTemplate} from "../interfaces";
+import PF2E from "./systems/pf2e";
 
 export default class LeapYear extends ConfigurationItemBase{
     /**
@@ -76,6 +77,7 @@ export default class LeapYear extends ConfigurationItemBase{
      * @param {number} year The year number to check
      */
     isLeapYear(year: number): boolean {
+        PF2E.checkLeapYearRules(this);
         if(this.rule === LeapYearRules.Gregorian){
             return year % 4 === 0 && (year % 100 !== 0 || (year % 100 === 0 && year % 400 === 0));
         } else if(this.rule === LeapYearRules.Custom){